+++ fabric/src/main/java/net/SpectrumFATM/black_archive/fabric/network/VortexTeleportPacket.java
@@ -3,10 +3,12 @@ package net.SpectrumFATM.black_archive.fabric.network;
 import net.SpectrumFATM.black_archive.fabric.BlackArchive;
 import net.SpectrumFATM.black_archive.fabric.config.BlackArchiveConfig;
 import net.SpectrumFATM.black_archive.fabric.sound.ModSounds;
+import net.SpectrumFATM.black_archive.fabric.util.SpaceTimeEventUtil;
 import net.SpectrumFATM.black_archive.fabric.util.WorldUtil;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.particle.ParticleTypes;
 import net.minecraft.registry.RegistryKey;
@@ -40,6 +42,8 @@ public class VortexTeleportPacket {
             String dimension = buf.readString(32767);
 
             server.execute(() -> {
+                SpaceTimeEventUtil.setComplexSpaceTimeEvent(player, true);
+                BlackArchive.LOGGER.info(String.valueOf(SpaceTimeEventUtil.isComplexSpaceTimeEvent(player)));
 
                 if (player.getItemCooldownManager().isCoolingDown(player.getMainHandStack().getItem())) {
                     player.sendMessage(Text.literal("You must wait before teleporting again.").formatted(Formatting.RED), false);
@@ -51,7 +55,6 @@ public class VortexTeleportPacket {
                 ServerWorld targetWorld = server.getWorld(dimensionKey);
 
                 if (targetWorld != null) {
-
                     player.getServerWorld().spawnParticles(ParticleTypes.SMOKE, player.getX(), player.getY(), player.getZ(), 10, 0.5, 0.5, 0.5, 0.0);
                     player.getWorld().playSound(null, player.getX(), player.getY(), player.getZ(), ModSounds.VORTEX_TP, SoundCategory.PLAYERS, 0.25f, 1.0f);