+++ fabric/src/main/java/net/SpectrumFATM/black_archive/fabric/tardis/upgrades/ModUpgrades.java
@@ -14,6 +14,7 @@ import java.util.Objects;
 public class ModUpgrades {
     public static RegistrySupplier<Upgrade> TELEPATHIC_UPGRADE;
     public static RegistrySupplier<Upgrade> REMOTE_UPGRADE;
+    public static RegistrySupplier<Upgrade> TEMPORAL_ORBIT_UPGRADE;
 
     static {
         REMOTE_UPGRADE = TRUpgrades.UPGRADE_DEFERRED_REGISTRY.register("remote_upgrade", () -> {
@@ -25,6 +26,11 @@ public class ModUpgrades {
             Item i = Items.PLAYER_HEAD;
             Objects.requireNonNull(i);
             return (new Upgrade(i::getDefaultStack, TRUpgrades.DIMENSION_TRAVEL, RegistryHelper.makeKey("telepathic_upgrade"), Upgrade.UpgradeType.SUB_UPGRADE)).setSkillPointsRequired(25).setPosition(7.0, 6.0);
+        });
+        TEMPORAL_ORBIT_UPGRADE = TRUpgrades.UPGRADE_DEFERRED_REGISTRY.register("temporal_orbit_upgrade", () -> {
+            Item i = Items.CLOCK;
+            Objects.requireNonNull(i);
+            return (new Upgrade(i::getDefaultStack, TRUpgrades.DIMENSION_TRAVEL, RegistryHelper.makeKey("temporal_orbit_upgrade"), Upgrade.UpgradeType.SUB_UPGRADE)).setSkillPointsRequired(25).setPosition(6.0, 5.0);
         });
     }